[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int a, b, result;
+long long a, b, result;
 
 int main()
 {
